Updated install.sh script

+++ install.sh
@@ -1,21 +1,55 @@
 #!/bin/bash
-echo "[+] Sit back and relax till this installation finishes"
+echo "[+] Sit back and relax till this installation finishes."
+
 sleep 2
-sudo apt install figlet;
-sudo apt install toilet;
-apt install golang;
-apt install nmap;
-apt install sublist3r
+
+if [ ! -f /bin/figlet ];then
+	sudo apt install figlet -y;
+fi
+
+if [ ! -f /bin/toilet ];then
+	sudo apt install toilet -y;
+fi
+
+if [ ! -f /bin/go ];then
+	sudo apt install golang -y;
+fi
+
+if [ ! -f /bin/nmap ];then
+	sudo apt install nmap -y;
+fi
+
+if [ ! -f /bin/sublist3r ];then
+	sudo apt install sublist3r -y;
+fi
+
 sleep 2
+
 echo "[+] Installing Go tools..."
-go install github.com/tomnomnom/assetfinder@latest
+if [ ! -f /bin/assetfinder ]; then
+	go install github.com/tomnomnom/assetfinder@latest
+fi
+
 sleep 2
-go install github.com/tomnomnom/httprobe@latest
+
+if [ ! -f /bin/httprobe ]; then
+	go install github.com/tomnomnom/httprobe@latest
+fi
+
 sleep 4
-go install -v github.com/projectdiscovery/httpx/cmd/httpx@latest
+
+if [ ! -f /bin/httpx ]; then
+	go install -v github.com/projectdiscovery/httpx/cmd/httpx@latest
+fi
+
 sleep 2
-echo "[+] Transferring the Go binaries to /usr/bin"
+echo "[+] Configuring requirements."
+
+#$HOME/go/bin
+sudo cp $HOME/go/bin/httprobe /usr/bin
+sudo cp $HOME/go/bin/assetfinder /usr/bin
+sudo cp $HOME/go/bin/httpx /usr/bin
 
 sleep 2
 
-echo "Done"
+echo "[+] Setup completed"
